First pass at implementing a typeclass-based `toRealObjSafe` function
This will eventually close the create-any-object-on-host-VM security hole as well as make result extraction more predictable (which currently is a problem where `toRealObj[Throwable]` appears to discard stack traces), at the cost of a small amount of boilerplate for each type you want to deserialize

Ported over `IOTest`, `MathTest`, and some internal calls to use this new mechanism; it seems to work. Long way to go to port the rest of the codebase.